services/observability/requirements.txt
kafka-python
influxdb-client
